Cleanup before implementing trending movie paging

+++ app/src/main/java/com/codepunk/moviepunk/data/remote/entity/RemoteMoviePage.kt
@@ -13,5 +13,5 @@ data class RemoteMoviePage(
     val totalPages: Int = 0,
 
     @SerialName(value = "total_results")
-    val totalResults: Int = 0
+    val totalResults: Long = 0
 )
